/products/category/1

+++ src/main/java/ec/edu/ups/icc/fundamentos01/products/dtos/UpdateProductDto.java
@@ -1,7 +1,5 @@
 package ec.edu.ups.icc.fundamentos01.products.dtos;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
@@ -22,7 +20,7 @@ public class UpdateProductDto {
 
     @NotNull(message = "El precio del producto es obligatorio")
     @DecimalMin(value = "0", inclusive = true, message = "El precio debe ser mayor o igual a 0")
-    private BigDecimal price;
+    private Double price;
 
     @NotNull(message = "El stock del producto es obligatorio")
     @Min(value = 0, message = "El stock debe ser mayor o igual a 0")
@@ -31,7 +29,7 @@ public class UpdateProductDto {
     public UpdateProductDto() {
     }
 
-    public UpdateProductDto(String name, String description, BigDecimal price, Integer stock) {
+    public UpdateProductDto(String name, String description, Double price, Integer stock) {
         this.name = name;
         this.description = description;
         this.price = price;
@@ -45,8 +43,8 @@ public class UpdateProductDto {
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
 
-    public BigDecimal getPrice() { return price; }
-    public void setPrice(BigDecimal price) { this.price = price; }
+    public Double getPrice() { return price; }
+    public void setPrice(Double price) { this.price = price; }
 
     public Integer getStock() { return stock; }
     public void setStock(Integer stock) { this.stock = stock; }
